Simple message deletion implemented (closes #29).
This serves not much of a big purpose, but anyhow Whatsapp does it.
Message is removed from current chat window and a question is aked
to confirm if message should also be deleted from the database.

If the answer is negative, the message will appear again when the
chat window is reloaded, so you can show a screenshot of a
conversation to a third party without sensitive information but
still keep it for the next time you open the chat.

If the message is enqueued it'd be a good idea to dequeue it, but
that is not implementedon this commit.